+++ forge/src/main/java/net/mehvahdjukaar/supplementaries/mixins/forge/ExplorationMapFunctionMixin.java
@@ -2,6 +2,7 @@ package net.mehvahdjukaar.supplementaries.mixins.forge;
 
 import net.mehvahdjukaar.supplementaries.configs.CommonConfigs;
 import net.mehvahdjukaar.supplementaries.integration.CompatHandler;
+import net.mehvahdjukaar.supplementaries.integration.QuarkCompat;
 import net.mehvahdjukaar.supplementaries.integration.forge.quark.CartographersQuillItem;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.ItemStack;
@@ -32,7 +33,7 @@ public abstract class ExplorationMapFunctionMixin {
     @Inject(method = "run", at = @At("HEAD"), cancellable = true)
     public void turnToQuill(ItemStack stack, LootContext context, CallbackInfoReturnable<ItemStack> cir){
         if(CompatHandler.QUARK && CommonConfigs.Tweaks.REPLACE_VANILLA_MAPS.get()){
-            cir.setReturnValue(CartographersQuillItem.forStructure(context.getLevel(), this.destination,
+            cir.setReturnValue(QuarkCompat.makeAdventurerQuill(context.getLevel(), this.destination,
                     this.searchRadius, this.skipKnownStructures, this.zoom, this.mapDecoration, null, 0));
         }
     }