+++ forge/src/main/java/net/mehvahdjukaar/supplementaries/mixins/forge/TreasureMapForEmeraldsMixin.java
@@ -49,7 +49,7 @@ public abstract class TreasureMapForEmeraldsMixin {
     public void turnToQuill(Entity trader, RandomSource random, CallbackInfoReturnable<MerchantOffer> cir) {
         if (trader.level instanceof ServerLevel serverLevel) {
             if (CompatHandler.QUARK && CommonConfigs.Tweaks.REPLACE_VANILLA_MAPS.get()) {
-                ItemStack map = QuarkCompatImpl.makeAdventurerQuill(serverLevel, this.destination,
+                ItemStack map = QuarkCompat.makeAdventurerQuill(serverLevel, this.destination,
                         100, true, 2, this.destinationType, null, 0);
                 map.setHoverName(Component.translatable(this.displayName));
                 int uses = CommonConfigs.Tweaks.QUILL_MAX_TRADES.get();